Move `AbstractionConstructor.BuildOutput` to `AbstractionConstructorTerm`
