+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/expressions/AbstractionConstructor.kt
@@ -71,72 +71,6 @@ abstract class AbstractionConstructor : FirstOrderExpression() {
         }
     }
 
-    data class BuildOutput(
-        val expressionLazy: Lazy<AbstractionConstructor>,
-        val argumentDeclarationBlockLazy: Lazy<StaticBlock>,
-    ) {
-        val expression: AbstractionConstructor by expressionLazy
-        val argumentDeclarationBlock: StaticBlock by argumentDeclarationBlockLazy
-    }
-
-    companion object {
-        fun build(
-            context: BuildContext,
-            term: AbstractionConstructorTerm,
-        ): BuildOutput {
-            val outerScope = context.outerScope
-
-            val argumentDeclarationBuildOutput = ArgumentDeclaration.build(
-                outerScope = outerScope,
-                argumentTypeTerm = term.argumentType,
-            )
-
-            val argumentDeclaration by argumentDeclarationBuildOutput.argumentDeclarationLazy
-            val argumentDeclarationBlock by argumentDeclarationBuildOutput.argumentDeclarationBlockLazy
-
-            val expressionStub = object : Stub<AbstractionConstructor> {
-                override val resolved: AbstractionConstructor by lazy {
-                    val declaredImageTypeBody = term.declaredImageType?.let {
-                        TypeExpression.build(
-                            outerScope = context.outerScope,
-                            term = it,
-                        ).resolved
-                    }
-
-                    val declaredImageTypeAnalysis = declaredImageTypeBody?.evaluateAsType()
-
-                    val innerScope = argumentDeclarationBlock.chainWith(
-                        outerScope = outerScope,
-                    )
-
-                    val image = build(
-                        context = BuildContext(
-                            outerScope = innerScope,
-                        ),
-                        term = term.image,
-                    )
-
-                    return@lazy object : AbstractionConstructor() {
-                        override val outerScope: StaticScope = context.outerScope
-
-                        override val term: ExpressionTerm = term
-
-                        override val argumentDeclaration = argumentDeclaration
-
-                        override val declaredImageType: TypeAlike? = declaredImageTypeAnalysis?.typeOrIllType
-
-                        override val image by lazy { image.resolved }
-                    }
-                }
-            }
-
-            return BuildOutput(
-                expressionLazy = expressionStub.asLazy(),
-                argumentDeclarationBlockLazy = argumentDeclarationBuildOutput.argumentDeclarationBlockLazy,
-            )
-        }
-    }
-
     override fun bindDirectly(dynamicScope: DynamicScope): Thunk<Value> = Thunk.pure(
         ComputableAbstraction(
             outerDynamicScope = dynamicScope,
