+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/expressions/Expression.kt
@@ -233,7 +233,7 @@ abstract class Expression {
             context: BuildContext,
             term: ExpressionTerm,
         ): Stub<Expression> = when (term) {
-            is AbstractionConstructorTerm -> AbstractionConstructor.build(
+            is AbstractionConstructorTerm -> AbstractionConstructorTerm.build(
                 context = context,
                 term = term,
             ).expressionLazy.asStub()
